Add type casting for c10::ArrayRef<double>
This patch adds pybind11 type casting to pass in arguments for
function `quad_pool2d`.

+++ include/torch_geopooling/quadpool.h
@@ -15,7 +15,7 @@ quad_pool2d(
     const torch::Tensor& tiles,
     const torch::Tensor& input,
     const torch::Tensor& weight,
-    const c10::ArrayRef<float>& exterior,
+    const c10::ArrayRef<double>& exterior,
     bool training = true,
     std::optional<std::size_t> max_depth = std::nullopt,
     std::optional<std::size_t> capacity = std::nullopt,
